requirements.txt
keras-contrib==2.0.8 # pip install git+https://www.github.com/keras-team/keras-contrib.git
torch==1.4.0 # pip install -f https://download.pytorch.org/whl/torch_stable.html torch==1.4.0
allennlp==0.9.0
python>=3.6
transformers>=2.8.0
gluonnlp==0.6.0
gensim==3.8.0
keras==2.2.4
keras-contrib==2.0.8 # pip install git+https://www.github.com/keras-team/keras-contrib.git
metric==0.10.0
nltk==3.4.5
pytorch-crf==0.7.2
scikit-learn==0.21.2
seqeval==0.0.12
tensorflow==1.14.0
tensorflow-gpu==1.14.0
tqdm==4.40.2
kobert-transformers==0.4.1
kss==1.3.1
msgpack==1.0.0
hgtk==0.1.3
python-docx==0.8.10
mecab-python # https://cleancode-ws.tistory.com/97 참조
konlpy==0.5.2